refactor: sub-work items components, hooks and types
