+++ packages/types/src/issues/issue_sub_issues.d.ts
@@ -20,3 +20,20 @@ export type TIssueSubIssuesStateDistributionMap = {
 export type TIssueSubIssuesIdMap = {
   [issue_id: string]: string[];
 };
+
+export type TSubIssueOperations = {
+  copyLink: (path: string) => void;
+  fetchSubIssues: (workspaceSlug: string, projectId: string, parentIssueId: string) => Promise<void>;
+  addSubIssue: (workspaceSlug: string, projectId: string, parentIssueId: string, issueIds: string[]) => Promise<void>;
+  updateSubIssue: (
+    workspaceSlug: string,
+    projectId: string,
+    parentIssueId: string,
+    issueId: string,
+    issueData: Partial<TIssue>,
+    oldIssue?: Partial<TIssue>,
+    fromModal?: boolean
+  ) => Promise<void>;
+  removeSubIssue: (workspaceSlug: string, projectId: string, parentIssueId: string, issueId: string) => Promise<void>;
+  deleteSubIssue: (workspaceSlug: string, projectId: string, parentIssueId: string, issueId: string) => Promise<void>;
+};
