+++ web/core/components/issues/issue-detail-widgets/relations/helper.tsx
@@ -1,18 +1,17 @@
 "use client";
 import { useMemo } from "react";
 import { usePathname } from "next/navigation";
+// plane imports
 import { EIssueServiceType, ISSUE_DELETED, ISSUE_UPDATED } from "@plane/constants";
 import { useTranslation } from "@plane/i18n";
 import { TIssue, TIssueServiceType } from "@plane/types";
 import { TOAST_TYPE, setToast } from "@plane/ui";
-// constants
-// helper
-import { copyTextToClipboard } from "@/helpers/string.helper";
+import { copyUrlToClipboard } from "@plane/utils";
 // hooks
 import { useEventTracker, useIssueDetail } from "@/hooks/store";
 
 export type TRelationIssueOperations = {
-  copyText: (text: string) => void;
+  copyLink: (path: string) => void;
   update: (workspaceSlug: string, projectId: string, issueId: string, data: Partial<TIssue>) => Promise<void>;
   remove: (workspaceSlug: string, projectId: string, issueId: string) => Promise<void>;
 };
@@ -29,9 +28,8 @@ export const useRelationOperations = (
 
   const issueOperations: TRelationIssueOperations = useMemo(
     () => ({
-      copyText: (text: string) => {
-        const originURL = typeof window !== "undefined" && window.location.origin ? window.location.origin : "";
-        copyTextToClipboard(`${originURL}${text}`).then(() => {
+      copyLink: (path) => {
+        copyUrlToClipboard(path).then(() => {
           setToast({
             type: TOAST_TYPE.SUCCESS,
             title: t("common.link_copied"),
@@ -39,7 +37,7 @@ export const useRelationOperations = (
           });
         });
       },
-      update: async (workspaceSlug: string, projectId: string, issueId: string, data: Partial<TIssue>) => {
+      update: async (workspaceSlug, projectId, issueId, data) => {
         try {
           await updateIssue(workspaceSlug, projectId, issueId, data);
           captureIssueEvent({
@@ -56,7 +54,7 @@ export const useRelationOperations = (
             type: TOAST_TYPE.SUCCESS,
             message: t("entity.update.success", { entity: entityName }),
           });
-        } catch (error) {
+        } catch {
           captureIssueEvent({
             eventName: ISSUE_UPDATED,
             payload: { state: "FAILED", element: "Issue detail page" },
@@ -73,7 +71,7 @@ export const useRelationOperations = (
           });
         }
       },
-      remove: async (workspaceSlug: string, projectId: string, issueId: string) => {
+      remove: async (workspaceSlug, projectId, issueId) => {
         try {
           return removeIssue(workspaceSlug, projectId, issueId).then(() => {
             captureIssueEvent({
@@ -82,7 +80,7 @@ export const useRelationOperations = (
               path: pathname,
             });
           });
-        } catch (error) {
+        } catch {
           captureIssueEvent({
             eventName: ISSUE_DELETED,
             payload: { id: issueId, state: "FAILED", element: "Issue detail page" },
@@ -91,7 +89,7 @@ export const useRelationOperations = (
         }
       },
     }),
-    [pathname, removeIssue, updateIssue]
+    [captureIssueEvent, entityName, pathname, removeIssue, t, updateIssue]
   );
 
   return issueOperations;
