+++ web/core/components/issues/issue-detail-widgets/sub-issues/root.tsx
@@ -21,7 +21,7 @@ export const SubIssuesCollapsible: FC<Props> = observer((props) => {
   const { workspaceSlug, projectId, issueId, disabled = false, issueServiceType } = props;
   // store hooks
   const { openWidgets, toggleOpenWidget } = useIssueDetail(issueServiceType);
-  // derived state
+  // derived values
   const isCollapsibleOpen = openWidgets.includes("sub-issues");
 
   return (
