--- web/core/components/issues/sub-issues/index.ts
@@ -1,1 +1,0 @@
-export * from "./root";
